Issue 38:	Implement payment - updated checkout (added shipping cost on all pages, used user friendly name for shipping method). Added transaction id to orders table.
git-svn-id: http://xe-shop.googlecode.com/svn/trunk@281 a237117d-c3a3-7e14-bcca-fd8395c59b2d

+++ plugins_payment/paypal_payments_standard/PaypalPaymentsStandard.php
@@ -21,49 +21,47 @@ class PaypalPaymentsStandard extends PaymentMethodAbstract
 
     public function onOrderConfirmationPageLoad($module_srl)
     {
-        // Check if cart doesn't exist anymore, which means the order has already been created
-        $cartRepo = new CartRepository();
-        $logged_info = Context::get('logged_info');
-        $cart = $cartRepo->getCart($module_srl, null, $logged_info->member_srl, session_id());
-        $orderRepository = new OrderRepository();
-        // TODO Check if cart exists and if order has not been already created
-        if(!$cart)
-        {
-            // TODO What to check for not logged in users?
-            $order = $orderRepository->getLastOrder($module_srl, $logged_info->member_srl);
-            Context::set('order_srl', $order->order_srl);
-            return;
-        }
-
         $tx_token = Context::get('tx');
         if(!$tx_token || !$this->pdt_token)
         {
-            throw new Exception("Invalid input. Please add transaction token and authorization token.");
+            return;
         }
 
-        $params = array();
-        $params['cmd'] = '_notify-synch';
-        $params['tx'] = $tx_token;
-        $params['at'] = $this->pdt_token;
-
-        $paypalAPI = new PaypalPaymentsStandardAPI();
-        $response = $paypalAPI->request(self::SANDBOX_URL, $params);
-        $response_array = explode("\n", $response);
-        if($response_array[0] == 'SUCCESS')
+        $orderRepository = new OrderRepository();
+        $order = $orderRepository->getOrderByTransactionId($tx_token);
+        if(!$order)
         {
-            $order = $orderRepository->getOrderFromCart($cart);
-            $order->save(); //obtain srl
-            $order->saveCartProducts($cart);
-            $cart->delete();
+            $cartRepo = new CartRepository();
+            $logged_info = Context::get('logged_info');
+            $cart = $cartRepo->getCart($module_srl, null, $logged_info->member_srl, session_id());
 
-            Context::set('order_srl', $order->order_srl);
-            // Override cart, otherwise it would still show up with products
-            Context::set('cart', null);
+            $params = array();
+            $params['cmd'] = '_notify-synch';
+            $params['tx'] = $tx_token;
+            $params['at'] = $this->pdt_token;
+
+            $paypalAPI = new PaypalPaymentsStandardAPI();
+            $response = $paypalAPI->request(self::SANDBOX_URL, $params);
+            $response_array = explode("\n", $response);
+            if($response_array[0] == 'SUCCESS')
+            {
+                $order = $orderRepository->getOrderFromCart($cart);
+                $order->save(); //obtain srl
+                $order->saveCartProducts($cart);
+                $cart->delete();
+
+                Context::set('order_srl', $order->order_srl);
+                // Override cart, otherwise it would still show up with products
+                Context::set('cart', null);
+            }
+            else
+            {
+                throw new Exception("There was some error from PDT");
+            }
         }
-        else
-        {
-            throw new Exception("There was some error from PDT");
-        }
+
+        Context::set('order_srl', $order->order_srl);
+        return;
     }
 }
